auth_service/requirements.txt
uvicorn==0.34.1
fastapi==0.115.12
httpx==0.28.1
pydantic==2.11.4
python_jose==3.4.0
SQLAlchemy==2.0.41
passlib==1.7.4
pika==1.3.2
PyJWT==2.10.1
redis==5.2.1
srp==1.0.22
aio-pika==9.5.5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncio-redis==0.16.0
python-multipart==0.0.20

